requirements.txt
web3
pyyaml
pytest